v2.8.5.0-Release Lots of changes, see changelog

+++ GagSpeak/UI/EquipmentAndCombos/GagLockFilterCombo.cs
@@ -32,7 +32,7 @@ public sealed class GagLockFilterCombo
     /// <item><c>label</c><param name="label"> - The label to display outside the combo box</param></item>
     /// <item><c>layerindex</c><param name="layerIndex"> - a list where the stored selection from the list is saved</param></item>
     /// </list> </summary>
-    public void Draw(int ID, CharacterHandler characterHandler, int layerIndex, int width) {
+    public void Draw(int ID, CharacterHandler characterHandler, int layerIndex, float width) {
         try
         {
             ImGui.SetNextItemWidth(width);
@@ -68,7 +68,7 @@ public sealed class GagLockFilterCombo
     /// <item><c>label</c><param name="label"> - The label to display outside the combo box</param></item>
     /// <item><c>layerindex</c><param name="layerIndex"> - a list where the stored selection from the list is saved</param></item>
     /// </list> </summary>
-    public void Draw(int ID, ref string label, CharacterHandler characterHandler, int whitelistIdx, int layerIndex, int width) {
+    public void Draw(int ID, ref string label, CharacterHandler characterHandler, int whitelistIdx, int layerIndex, float width) {
         try
         {
             // set the next item width to the width we want
